[Revision] Documentation
Went back and added javadocs for all sprint 2 classes.
