+++ Utilities/CharacterCreator/CharacterCreator.java
@@ -6,6 +6,12 @@ import Player.StatsBuilder;
 import Utilities.Die;
 import Utilities.NameGenerator.NameGenerator;
 
+/**
+ * The CharacterCreator class is responsible for creating new player characters with random names and stats.
+ * The creator has the functionality of allowing the player to reroll indefinitely.
+ * 
+ * @version 1.0
+ */
 public class CharacterCreator {
     private Player createdPlayer;
     public CharacterCreator() {
@@ -13,6 +19,11 @@ public class CharacterCreator {
         this.createdPlayer = new Player(this.rerollName(), this.rerollStats());
     }
 
+    /**
+     * Rerolls the stat attributes of the player. Uses the die class to randomize stat generation.
+     * 
+     * @return randomized stats object
+     */
     public PlayerStats rerollStats() {
         Die die = new Die();
         StatsBuilder statsBuilder = new StatsBuilder();
@@ -21,10 +32,19 @@ public class CharacterCreator {
         return stats;
     }
 
+    /**
+     * Returns the rerolled player.
+     * @return Player object
+     */
     public Player getPlayer() {
         return this.createdPlayer;
     }
 
+    /**
+     * Generates a random name with a preset of prefixes, suffixes. Uses the NameGenerator class.
+     * 
+     * @return randomized name
+     */
     public String rerollName() {
         System.out.println("Assigning random name...");
         String characterName = NameGenerator.generateName();
