+++ Utilities/DataStructures/Heap.java
@@ -1,0 +1,5 @@
+package Utilities.DataStructures;
+
+public interface Heap {
+    
+}
