+++ Utilities/ViewHelper.java
@@ -4,7 +4,17 @@ import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
 
+/**
+ * ViewHelper is responsible for view management within the client.
+ * Some example uses are to print consistent headers, print separators
+ * and print the starting game logo.
+ */
 public class ViewHelper {
+
+    /**
+     * Prints the desired amount of dash separators.
+     * @param n number of dashes in separator
+     */
     public static void printSeparator(int n) {
         for (int i = 0; i < n; i++) {
             System.out.print("-");
@@ -12,12 +22,19 @@ public class ViewHelper {
         System.out.println();
     }
 
+    /**
+     * Prints header with desired title.
+     * @param title to be printed as header
+     */
     public static void printHeader(String title) {
         printSeparator(25);
         System.out.println(title);
         printSeparator(25);
     }
 
+    /**
+     * Prints the SnSLogo.
+     */
     public static void printLogo() {
         //Resource: https://stackoverflow.com/questions/15695984/java-print-contents-of-text-file-to-screen
         try {
@@ -32,6 +49,9 @@ public class ViewHelper {
         }
     }
 
+    /**
+     * Prints out the main menu with game logic.
+     */
     public static void printMainMenu() {
         System.out.println("Please select an option to continue.");
         System.out.println("1. Reroll Character.");
